requirements.txt
Flask==0.10.1
RPi.Gpio
requests